+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [3.0.4]
+
+- Fix bug where gemini models would add code block artifacts to the end of text content
+- Fix context mention menu visual issues on light themes
+
 ## [3.0.2]
 
 - Adds block anchor matching for more reliable diff edits (if 3+ lines, first and last line are used as anchors to search for)